feat: splitted generateDocs and changed the output
Splitted the `generateDocs` function in `extract` and `organise`
functions.
The `extract` function extracts comments from the code and it cleans
them up.
The `organise` function returns all the sections and the names of the
root sections. Every returned section has a `subsections` property which
lists the names of the subsections.

BREAKING CHANGE